# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Quan Li Lich Phong Lab",
        "version": "1.0.0",
        "title": "Lab Calender Server"
    },
    "schemes": {
        "http": "http",
        "https": "https"
    },
    "securityDefinitions": {
        "bearerAuth": {
            "name": "Authorization",
            "in": "header",
            "type": "apiKey",
            "description": "Bearer {token}"
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/user/sendMail":{
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Send mail to user",
                "description": "Gửi mail cho người dùng",
                "operationId": "sendMail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Send Mail",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SendMail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/SendMail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        
        },
...
        "/api/user/getUserById/{uid}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Tìm user với mã UID",
                "description": "Trả về 1 người dùng và thông tin của người dùng",
                "operationId": "getUserById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Mã uid của user",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Invalid UID supplied"
                    },
                    "404": {
                        "description": "USER NOT FOUND"
                    }
                }
            }
        },
...
        "api/user/delete/{uid}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Xóa người dùng",
                "description": "Chỉ admin mới có thể xóa người dùng",
                "operationId": "deleteUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Mã uid của user",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Invalid UID supplied"
                    },
                    "404": {
                        "description": "USER NOT FOUND"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "SendMail": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email của người dùng cần gửi",
                    "required": true,
                    "example": "toExample@gmail.com"
                },
                "subject": {
                    "type": "string",
                    "format": "string",
                    "description": "Tiêu đề của email",
                    "required": true,
                    "example": "Tiêu đề của email"
                },
                "text": {
                    "type": "string",
                    "format": "string",
                    "description": "Nội dung của email",
                    "required": true,
                    "example": "Nội dung của email"
                },
                "html": {
                    "type": "string",
                    "format": "string",
                    "description": "Nội dung của email",
                    "required": true,
                    "example": "Nội dung của email"
                }
            }
        },
...
    }
}